/home/fox/release/dtrace/dtrace-20120131.tar.bz2

+++ driver/systrace.c
@@ -2129,7 +2129,7 @@ static void *sys_seq_start(struct seq_file *seq, loff_t *pos)
 {
 	if (*pos >= NSYSCALL + NSYSCALL32)
 		return 0;
-	return (void *) (*pos + 1);
+	return (void *) ((int) *pos + 1);
 }
 static void *sys_seq_next(struct seq_file *seq, void *v, loff_t *pos)
 {	long	n = (long) v;
